Added exception Handling.

+++ src/main/java/com/example/management_system/services/EmployeeServiceImpl.java
@@ -1,7 +1,12 @@
 package com.example.management_system.services;
 
+import com.example.management_system.converter.EmployeeConverter;
 import com.example.management_system.dao.EmployeeDao;
+import com.example.management_system.dto.EmployeeDto;
+import com.example.management_system.dto.EmployeeSubDto;
 import com.example.management_system.entities.Employee;
+import com.example.management_system.exception.ResourceAlreadyExistException;
+import com.example.management_system.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -9,23 +14,25 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class EmployeeServiceImpl implements EmployeeService {
 
     @Autowired
     private EmployeeDao employeeDao;
+    private EmployeeConverter employeeConverter = new EmployeeConverter();
 
     @Override
-    public List<Employee> getAll() {
-        return employeeDao.findAll();
+    public List<EmployeeSubDto> getAll() {
+        if(employeeDao.findAll().isEmpty()){
+            throw new ResourceNotFoundException("Employees");
+        }
+        return employeeConverter.entityToDto(employeeDao.findAll());
     }
 
     @Override
-    public List<Employee> getAllByPageSort(int pageNumber, int pageSize, String sortBy, String sortDirection){
+    public List<EmployeeSubDto> getAllByPageSort(int pageNumber, int pageSize, String sortBy, String sortDirection){
 
         Sort sort = null;
         if(sortDirection.equalsIgnoreCase("asc")){
@@ -37,28 +44,72 @@ public class EmployeeServiceImpl implements EmployeeService {
         Pageable p = PageRequest.of(pageNumber, pageSize, sort);
         Page<Employee> employeePage = employeeDao.findAll(p);
         List<Employee> allEmployee = employeePage.getContent();
-        return allEmployee;
+        if(allEmployee.isEmpty()){
+            throw new ResourceNotFoundException("Employees");
+        }
+        return employeeConverter.entityToDto(allEmployee);
     }
 
     @Override
-    public Optional<Employee> getById(long id) {
-        return employeeDao.findById(id);
+    public EmployeeSubDto getById(long id){
+        if(employeeDao.findById(id) == null){
+            throw new ResourceNotFoundException("Employee", "id", id);
+        }
+        return employeeConverter.entityToDto(employeeDao.findById(id));
+    }
+    
+    @Override
+    public List<EmployeeSubDto> getByNameAndAge(String name, int age){
+        if(employeeDao.findByNameAndAge(name, age).isEmpty()){
+            throw new ResourceNotFoundException("Employees");
+        }
+        return employeeConverter.entityToDto(employeeDao.findByNameAndAge(name, age));
+    }
+
+    public List<EmployeeSubDto> getBySalary(long salary){
+        if(employeeDao.findBySalary(salary).isEmpty()){
+            throw new ResourceNotFoundException("Employees");
+        }
+        return employeeConverter.entityToDto(employeeDao.findBySalary(salary));
     }
 
     @Override
-    public Employee add(Employee employee) {
+    public List<EmployeeSubDto> getByPosition(String position){
+        if(employeeDao.findByPosition(position).isEmpty()){
+            throw new ResourceNotFoundException("Employees");
+        }
+        return employeeConverter.entityToDto(employeeDao.findByPosition(position));
+    }
+
+    @Override
+    public List<EmployeeSubDto> getByPositionAndDepartment(String position, String department){
+        if(employeeDao.findByPositionAndDepartment(position, department).isEmpty()){
+            throw new ResourceNotFoundException("Employees");
+        }
+        return employeeConverter.entityToDto(employeeDao.findByPositionAndDepartment(position, department));
+    }
+    @Override
+    public Employee add(EmployeeDto employeeDto) throws ResourceAlreadyExistException {
+
+        if(employeeDao.findByEmail(employeeDto.getEmail()) != null){
+            throw new ResourceAlreadyExistException("Employee", "email", employeeDto.getEmail());
+        }
+        Employee employee = employeeConverter.dtoToEntity(employeeDto);
         employeeDao.save(employee);
         return employee;
     }
 
     @Override
-    public String update(Employee employee) {
+    public String update(Employee employee){
         employeeDao.save(employee);
         return "Employee Details Updated";
     }
 
     @Override
     public String deleteById(long id) {
+        if(employeeDao.findById(id) == null){
+            throw new ResourceNotFoundException("Employee", "id", id);
+        }
         employeeDao.deleteById(id);
         return "Deleted Successfully";
     }
@@ -68,4 +119,7 @@ public class EmployeeServiceImpl implements EmployeeService {
         employeeDao.deleteAll();
         return "All records deleted";
     }
+
+
+
 }
